Fix renderer updatnig pages too early
Player EVT_ZONE triggers before we add new pages (neighbours) to new curPage (loaded from server). Because of this the renderer doesn't have new pages loaded, so neighbour pages are null and unrendered). We could postpone updating pages, but the problem is we need to know that the current page HAS changed so that we correctly modify our offset and neighbours. So, instead we can update pages twice: once (immediately) when our curPage changes (locally hit EVT_ZONE), and again when our neighbour pages have loaded from the server